requirements.txt
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.11
django-rest-swagger==2.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
PyJWT==1.6.1
pytz==2017.3
requests==2.18.4
simplejson==3.13.2
uritemplate==3.0.0
urllib3==1.22
